change long to BigDecimal

+++ match/src/main/java/com/itranswarp/crypto/match/MatchRecord.java
@@ -1,13 +1,15 @@
 package com.itranswarp.crypto.match;
 
+import java.math.BigDecimal;
+
 public class MatchRecord {
 
 	public final long takerOrderId;
 	public final long makerOrderId;
-	public final long matchPrice;
-	public final long matchAmount;
+	public final BigDecimal matchPrice;
+	public final BigDecimal matchAmount;
 
-	public MatchRecord(long takerOrderId, long makerOrderId, long matchPrice, long matchAmount) {
+	public MatchRecord(long takerOrderId, long makerOrderId, BigDecimal matchPrice, BigDecimal matchAmount) {
 		this.takerOrderId = takerOrderId;
 		this.makerOrderId = makerOrderId;
 		this.matchPrice = matchPrice;
